[Tools] Move most of the generate_gclient-xwalk's logic to DEPS.xwalk.
Most of the code in generate_gclient-xwalk.py should not exist: it
functioned mostly as a translator that read values from DEPS.xwalk and
changed them into a format suitable for gclient.

We can get rid of this intermediate step by writing DEPS.xwalk in a
gclient-compatible format. In practice, this means that DEPS.xwalk is
now responsible for:

* Setting git revision values for repositories we track, such as
  chromium-crosswalk and ozone-wayland (as before).

* Automatically ignoring some directories in the checkout.

* Setting variables parsed by gclient, such as |solutions|.

All this also allows us to remove a lot of validation code that was
scattered around generate_gclient-xwalk.py.

generate_gclient-xwalk.py now exists merely for compatibility, as it
still needs to parse the XWALK_OS_ANDROID environment variable and the
--cache-dir option and append a few more lines to .gclient-xwalk
accordingly. I can see these features being moved somewhere else and the
script getting removed altogether in the future.

This is a significant step in the preparations for XWALK-1894.