DEPS.xwalk: Start tracking all Chromium dependencies from git.
Chromium is soon going to switch its canonical source repository to git
and we would have to follow suit anyway. In addition, this makes it
easier to cache directories across machines and, in some cases, the
network communication is faster and more reliable when using git.

Most of the work was already done in previous commits that moved the
majority of the logic in generate_gclient-xwalk.py to DEPS.xwalk itself.

We now simply have to adjust the solutions set in DEPS.xwalk, so that
instead of checking out a Chromium release directory from SVN and
pulling all Chromium dependencies from its DEPS file (which points to
SVN repositories for everything) we point directly to chromium-crosswalk
(which was recently adjusted to contain the commit adjusting DEPS and
.DEPS.git for a certain Chromium release) and its .DEPS.git.

The only unusual part of this commit is the moving of Ozone-Wayland to a
separate gclient solution: we need to do that because it has a .DEPS.git
which overrides gclient's recursion limit and would cause problems
because it too would try fetching Chromium. We explicitly set
|deps_file| in this solution to "" to prevent that.

BUG=XWALK-1894

(cherry picked from commit 4156cd875081ab00e30517578f77370dcee2ec49)

Conflicts:
	DEPS.xwalk