[Tools] Use .gclient's value of |cache_dir| in .gclient-xwalk.
The previous approach was not easy to use at all, as one had to manually
call `gclient sync' with --nohooks, call generate_gclient-xwalk.py
manually with --cache-dir and then run the other hooks.

Instead, just use the same value as .gclient so that all repositories
are cached in the same location, thus also reducing the differences
between .gclient and .gclient-xwalk.

--cache-dir is still accepted and used if passed for compatibility
reasons; support for it can be removed altogether in the next major
version (Crosswalk 8).

Related to XWALK-1184 and XWALK-1894.