[Android] Mark content as a local scheme in blink
Now there is no option to enable the cross-origin-access in the
context of content url in embedded mode. Regarding file scheme we
use setAllowUniversalAccessFromFileURLs option to support the
cross-orign-access. This way works for all the schemes which are
recognized as the "local" ones in blink. That's straightforward
to let content scheme follow this way to support the
cross-orgin-access setting.

BUG=XWALK-1959