[Android] Adjust annotation of Embedding API for documentation
Before reflection layer for shared mode lands, there are still
public methods in org.xwalk.core which are not supposed to be
Embedding API. So we still need to support doclava for now.

This commit includes following changes:
1, Reorder the declare sequence to make Exposed APIs prior to the
internal one with the same function name (which takes internal
class as paramter).
2. Add @since tag for all APIs, although since tag is not supported
in doclava, it's still worth to do it for API maintenance. This
also applies to the internal classes, before in future, the annotation
in org.xwalk.core will be generated from org.xwalk.core.internal by
script. So for now, we need to keep them synced.
3. Add @hide to the non-Embedding API public methods.

Related Bugs: XWALK-1856, XWALK-2017