[Android] Prompt dialog when CPU architecture mismatch
Previous commit c86dcf7e fixed running ARM on x86.
It was broken, due to the MACRO used to detect CPU
architecture now needs including a header file.

For the opposite, if installing x86 crosswalk on ARM,
the native library will not be available at all.
(ARM native library could still be run on x86 devices, but slow).
A ProcessInit exception will be thrown instead of an
UnsatisfiedLinkError which is expected in previous commit.
To fix this, try to find the UnsatisfiedLinkError in the cause
chain of the exception met while initiating crosswalk.

BUG=https://crosswalk-project.org/jira/browse/XWALK-2011