Move to M37.0.2062.20
- Skip SIMD command line for now.
- sync gyp_xwalk with gyp_chromium (r274762).
- Fixed the code for the Geolocation. Just like in M36 the delegate
has been refactored. It's now not a separate class to inherit from
but a regular object to instanciate. On other platforms than Tizen and Android
we need to make sure we return with the callback otherwise it hangs waiting for a reply.
As we do not support this on desktop we can just early return false.
- Fix 'screen' property not being replaced in a test : Blink commit 5fdac279
made the 'screen' property non-replaceable, so the only way to overwrite it
now is using v8tools utility to force.
- [Android] Strings now are generated differently through a dedicated GYP target.
- DEPS.xwalk: Start tracking Ozone-Wayland's Milestone-Summer branch.
- [Android] Fix resource packaging : In upstream the resource processing changed a
lot.
-[Tizen] Temporarily disable ALSA support : https://codereview.chromium.org/264973012 introduces a dependency on
ALSA's sequencing interface, which is currently disabled on Tizen.
- [Tizen] Add a patch for angle specific to Tizen to skip the .git/index dependency wrongly added. Fixed upstream
already http://code.google.com/p/chromium/issues/detail?id=395705
[Tizen] Temporarily disable WebRTC support (XWALK-2160)