Fix the invalid product string of Crosswalk.
The product string value should align with the value in user agent, which
will also be used by the DevTools ('Browser' section of version command).

The product string is meanful for remote debugging, eg: if the version in
product string is higer than the version of desktop Chrome. Then the
'chrome://inspect' UI should report the following warning string:
"You may need a newer version of desktop Chrome. Please try Chrome 36.0.1985.18 or later."

This is useful information for remote debugging, to make it work with non-Chrome
browsers (include Crosswalk) in device, the upstream patch under review should also
be required:
   https://codereview.chromium.org/367083002/

BUG=