[Tizen] Application resource protection implementation This is the first commit which includes: - Add the attribute "encryption" in the tizen:setting in config.xml. - If the value is "enable", the WRT must encrypt the HTML, CSS, JS file resources when the application is installed. And the WRT should be able to decrypt the encrypted data when the application is loaded for running. According to the spec, the encryption is performed at the installation moment (not at packaging time).
For security consideration, I do not implement the encyption & decryption algorithm. They need supports from the system level, such as the trust-zone on ARM or chaabi on IA. But at this moment the Tizen does not include such the security module. Another solution may be storing the security key within system-level storage like gnome-keyring or some external / internal storage. However this is still not supported by Tizen. So at present I leave the encryption & decryption not implemented until we get support from system in the future. I will be back to finish the feature when the security module is available at system-level.

BUG=XWALK-1172