[Extension] Implemented bulk data transmission enhancement feature.
Inspired by Thiago Marcos P. Santos <tmpsantos@gmail.com>

Enhanced the feature of postMessage from extension to renderer process
by introduced shared memory when transmitting big bulk of data, also
used a threshold kInlineMessageMaxSize to control if we need to use
shared memory or IPC message.

If the message to post exceeds the threshold, we will create a shared
memory to wrap the message and post the shared memory handle instead;
in the receiver process we'll restore the message from the received
shared memory and process it as before.

BUG: https://crosswalk-project.org/jira/browse/XWALK-1377