runtime: Add aura-only native support
This CL brings aura-only native support for the browser runtime, where views
toolkit dependency is opt out. This is particularly useful for embedded and
limited systems, where the runtime is used always as a fullscreen application
to the host window system. This is aimed to work with Ozone DRI or GBM
implementations.

Testing was performed on Ubuntu Linux 14.04 platform.

TEST= $ python xwalk/gyp_xwalk -Dchromeos=0 -Duse_ozone=1 -Dembedded=1 -Duse_udev=1 -Dozone_platform_gbm=1 -Dozone_platform_wayland=0 -Ddisable_nacl=1
      $ ninja -j16 -C out/Release xwalk
      $ ./out/Release/xwalk --ozone-platform=gbm http://hexgl.bkcore.com/play/