[Android] make_apk: Stop passing icon_dict everywhere.
Add a small amount of sanity to make_apk's code and consolidate the
value of the "icon" option/property in a single place, AppInfo.

Instead of keeping two separate variables, one in AppInfo with the value
passed to --icon and a separate one with the value from the manifest,
just set AppInfo's |icon| member with --icon if it is set, and "icon"
from the manifest otherwise (as the current logic does).