[Linux] Remove XWalkRuntimeTest.LoadGeolocationPage
Geolocation handler is not implemented on platforms other than Android
or Tizen therefore the test does not actually verify the feature. It
was added a year ago to make sure Crosswalk doesn't crash if there is
no geolocation provider. At the time the API was made to return a pointer
of the provider and as Linux was not creating one it was crashing. With
M37 the way geolocation requests work have been refactored with a callback
based system which returns always false in our implementation. We should
re-add a modified version of this test if we support that feature later
in Linux.

BUG=XWALK-2176