[Tizen] hardcode OZONE_WAYLAND_USE_XDG_SHELL env in service file
This change tells ozone wayland backend
that shell is supporting unstable xdg-shell.

This is needed for by tizen's Application.hide().

Alternative would be store this var in system wide env
ie: /etc/profile/crosswalk.sh
And eventually force a login shell (sh -l -c "...")
in this service file.

But for performance concern it's hardcoded there,
this change may be reverted later
if Tizen sessions decide to use login shells then it will work same.

Refer to this page for xdg-shell integration in Tizen :

https://wiki.tizen.org/wiki/Wayland_xdg-shell_protocol