[REG] Fixed regression error that NaCl module failed to load
When rolling Chromium to version 37.0.2062.76, the interface
content::ContentMainDelegate::ZygoteStarting has been changed,
which makes it fails to call the old implementation in
xwalk::XWalkMainDelegate. Reimplemented it as indicated by chrome.

Also according to patch 1d7be68, nacl::NaClHelper is introduced
with the refactor of MessageChannel to control the in-process
and out-of-process message queue behavior, which should be created
inside xwalk::XWalkContentRendererClient::RenderFrameCreated when
NaCl is enabled.

BUG=https://crosswalk-project.org/jira/browse/XWALK-2297