[Android] Rewrite generate_xwalk_core_library_aar.py.
The previous version was horribly bloated and broken. The main
issue (which is blocking XWALK-1930) is that CopyBinaries() recurses
into some directories and tries to copy each subpath individually. Not
only is this entirely unnecessary, but the code does it the wrong way
and, for ARM builds, creates something like
'jni/xwalk_core_library_java.jar/armeabi-v7a/libxwalkcore.so'.

Instead of just fixing one issue, rewrite the whole script in a saner,
much smaller and more Pythonic way:

* Remove a lot of unnecessary functions.
* Get rid of the "-s" option which was not being used at all.
* Stop calling sys.exit() in main() since it's supposed to return an
  exit code instead.
* Stop copying files from xwalk_core_library/ into an intermediate
  directory and just write them to the zip file directly.