[Android] Fix the issue about onJavascriptModalDialog was not called.
This patch is to fix the issue about onJavascriptModalDialog was not called
when this function was overridden in class of XWalkUIClient.
Modify the condition for override function, if parameter is not the
instance of specified, create a new instance.

BUG=XWALK-2534
(cherry picked from commit f4fad94a1612e927cd586d68918614b2dc6e468d)