[Android] Detect device cpu abi more precisely
With latest Houdini, if an app contains arm binaries,
the app will see system abi as if it's running on
an arm device. In detail, android.os.Build.CPU_ABI will
be "armeabli-v7a", as well as System.getProperty("os.arch").

The new approach is to read output of "getprop" command
out of process.

BUG=https://crosswalk-project.org/jira/browse/XWALK-2653