[Android] Try to use shared mode when CPU arch mismatches
Shared mode for crosswalk is sharing java classes and
native libraries. Previously, it will try to find library
only if internal java classes are not found.
This commit extends the condition to handle CPU arch
mismatches. In that case, java classes are available but
native library is not.

Implement this by move the loading of native library ahead
to the initialization of reflection layer.

BUG=https://crosswalk-project.org/jira/browse/XWALK-2633