[Android]  Allow friendly process to authenticate
with the DevTools server

Fixed the second issue with authentication during opening of
Chrome Dev Tools socket by the neighbour process on Android. The
first issue was fixed by commit cc7f50e

The usage model when neighbour process connects to the Chrome Dev Tools
socket is prohibited by Chromium, but for Crosswalk we have such usage
model and we have an ability to register uid of the neighbour if it calls
appropriate Crosswalk API which finally calls
XWalkDevToolsServer::AllowConnectionFromUid.

The change includes passing of XWalkDevToolsServer pointer to
UnixDomainServerSocketFactory and registering of
XWalkDevToolsServer::CanUserConnectToDevTools as a function for authentication
later on in the UnixDomainServerSocketFactory::Create.
Without this param xwalk tried to check authentication only from Chromium
while we need to check also vs uid registered through
XWalkDevToolsServer::AllowConnectionFromUid.