[TIZEN] Add media class property in manifest for audio streams
This change extends the manifest with 'xwalk_media_app_class' property that
can be used to tag audio streams in pulseaudio and Murphy, so that the Murphy
resource policy daemon can dictate when the audio will be paused and resumed
in response to other audio playing.

This depends on chromium-crosswalk commit 52d4347, so roll
chromium-crosswalk (88fc58a -> 52d4347).

BUG=XWALK-2788