[Android] Allow web socket connection from any uid for debugging
There is already a switch for remote debugging.
And allow uid to access is very unintuitive since:
1. uid for Android app is determined at install time, so it must be
   a dynamical setting for the allow uid.
2. There can't be "secure" rule for xwalk core to decide which
   uid to allow. So basically we need to allow any uid who requires
   if we still be using such mechanism for authentication.
So just directly allow web socket connection from any uid if
remote debugging is turned on.

Besides, previous uid access is put in the Security scenario DEFAULT,
which is not appropriate. Now create a new value ALLOW_SOCKET_ACCESS
for it.