gyp: Stop using Chromium's bundled freetype2.
Remove the |use_custom_freetype| variable from xwalk.gyp and use the
system's Freetype instead.

The use_custom_freetype code came in from content_shell's gyp code when
we started the project. It is only used there for DumpRenderTree,
requires additional code to be built and is not used by the Chromium
browser at all.

BUG=XWALK-1985
BUG=XWALK-2571