[Tizen] Remove Blink-Add-GCC-flag-Wno-narrowing-fix-64bits-build.patch.
Since b089b656 ("[Tizen] Stop unconditionally passing -Wall to the
compiler") we have stopped making the compiler generate more error
messages than intended. As that commit explains, Chromium has its own
mechanisms for deciding which parts of the code need -Wall and which
need it to be left out (since several pieces are built with -Werror),
and we should trust it to do the right thing.

This allows us to get rid of this patch since the code now builds
cleanly on both 32- and 64-bit architectures.

BUG=XWALK-917