[Tizen] Screen orientation lock should not depend on sensor.
If JS client invokes the API screen.orientation.lock, the actual screen should be locked to correct orientation no matter whether the sensor is connected. The CL is to remove the logic of checking sensor connection.

This CL also fixed the rotation bug. For landscape devices which have landscape-primary as default orientation, the rotation conversion ((rot - 1) % 4) gets invalid value that might cause unexpected result. This fix corrected the conversion in order to get valid rotation.