[Tizen] packaging: Always set the file permissions when calling install(1).
By default, install(1) installs everything with a permission of 0755,
which we do not need for anything but executables. Everything else is
installed with chmod 0644.

Be explicit and specify the permissions we want when installing all
files in the %install phase.

While here, regroup some calls to install(1) so that only binaries are
installed in the section we call "Binaries".