[Android] Fix the access issues from iframe.
This patch is to fix the cross origin issues when access the different
origin urls in iframe.
The instance of XWalkSettings was not initialized when the preference
value was got, so can not set the value to XWalkSettings.
Adjust the code to generate the instance of XWalkSettings before load
the preference value.

BUG=XWALK-2685, XWALK-2905
(cherry picked from commit e7280020888083b4fd88e737164985ad386a0404)