[Tizen] packaging: Use the system's protobuf instead of Chromium's.
protobuf is quite resource-heavy to compile, so using a version already
installed in the system makes the build faster and also reduces our
usage of bundled Chromium libraries.

The version in src/third_party/protobuf is similar to the one in
Tizen (which was recently updated to 2.6.1): 2.5.0 plus a lot of MIPS64
and aarch64 patches that made it into the 2.6 release series.