[Android] Fix file://*/*.xhtml will be opened in default browser.
Because the mime type of xhtml file is "application/xhtml+xml".
It's intercepted by XWalkNavigationHandler to not handle in Crosswalk.
The default behavior for xhtml should be the same as html file, that
is loaded in the Crosswalk no matter with file schema or http schema.

This commit is backported from 6056a6f27

BUG=XWALK-2912