Revert "[Android] Enable Experimental WebPlatform Features"
Enabling the entire set of web platform features currently marked
experimental in Blink is dangerous. The features in this set changes
across Chromium releases, and it can be the case that some of them just
are not ready for production at all.

Case in point: this enables support for the @viewport CSS rule, which
currently has problems and causes regressions for our users (see
XWALK-3574).

If we are interested in enabling certain features (the commit message of
the change being reverted mentions CSS Animations, for example), they
should be enabled individually.

This reverts commit 9255613355cf56547953f2ac816fbc4581659a49.

BUG=XWALK-322
BUG=XWALK-3574