Support to replace user agent string for XPK package
The Crosswalk web runtime should support to replace the user agent string when a
web application running on top of it, so that the remote server could return the
right response to the web application as expecting. The patch will enable this
feature by adding the 'xwalk_user_agent' field in manifest.json file on all
platforms except the Android.

BUG=XWALK-3689