[Android] Fix the display for the website's title after navigation.
This patch is to display the website's title when open multiple URLs and
press the back or forward button.
The event of update title will be fired in render frame host and title will be
compared to the one in navigation entry, if they are same, the update
title event will not be sent to Java layer.

BUG=XWALK-2333