[Android] make_apk: Provide better explanations for the version options.
Improve the descriptions of the "--app-version", "--app-versionCode" and
"--app-versionCodeBase" options. We start from the assumption that users
are already familiar with the Android App Manifest XML file, and then
explicitly mention how those three options map to the
"android:versionName" and "android:versionCode" attributes there, in
addition to providing examples of what valid arguments actually look
like for those options.

BUG=XWALK-3473