Fix compilation on Windows 64 bits
Code usually uses std functions which returns a size_t and size_t
is not going to fit into an int on 64 bits. Today it triggers a
warning of potential data loss.

We could disable the warning but it's better to fix the code.
Please note that we have various casts, these are mandatory (though
unfortunate) at the moment because the API in Chromium requires
explicitely an int for example. This can be fixed upstream and
bugs have been reported (e.g. crbug.com/167187).