[Android] make_apk: Stop suggesting architectures without downloads.
Even though we do not provide official Crosswalk downloads for ARM64 and
x86_64, we were still nagging users to provide APKs for both
architectures whenever they used make_apk.py in embedded mode.

Stop doing that and only suggest the architectures for which we provide
official downloads (ARM and x86). The code supporting ARM64 and x86_64
remains in place, and we do allow users to specify them in the
command-line if they have build Crosswalk for those architectures
themselves.

While here, reword and improve the message listing all the APKs that
have been created by the make_apk invocation.