Add a new hook to create Google Play Services's directory structure.
Ever since M42, building Chromium (and consequently Crosswalk) for
Android has required installing the Google Play Services library via the
`android` tool first. However, doing so involves manually accepting an
End User License Agreement (EULA), which makes things very difficult to
automate and also breaks our checkout+configuration process for users
who are building Crosswalk for the first time.

Since that library is actually needed only by Chromium but not
Crosswalk, we have been working around this requirement. When Crosswalk
started tracking Chromium M42, we added a few changes to
chormium-crosswalk. One of them was making gyp not fail if some
directories created by the Google Play Services library were not
present.

This solution needs to be rethought for the upcoming M43, as there is
now a build system check in the android_tools repository to fail if a
certain file installed by that library is not found.

The script being added by this change is the answer to this, and is
being landed before M43 itself because it also works independently of
it. If the android_tools repository is checked out, it tries to create
the Google Play Services directories gyp looks for as well as an empty
file with the name that the android_tools build system checks. This way,
we get rid of the corresponding change in chromium-crosswalk, prepare
for the upcoming change in android_tools and make most of these things
be controlled from the crosswalk repository itself.

BUG=XWALK-3955