[Android] Provide asynchronous interface for initializing the Crosswalk Library
The XWalkView's environment is initialized at the moment first XWalkView
object is instanciated. The initialization is kind of a heavy task which
includes loading native library, staring processes and so on, may cause
the UI thread freezed for a moment.

So this patch provides an interface for initializing the environment in
background and notify the host application once ready. This patch also
does the following things:

- Dismisses XWalkLibraryListener, use new XWalkLibraryInterface
- Adds new error type for the case that the CPU architecture
  mismatches. Removes unused code with same function from XWalkViewInternal
- Makes the init method of XWalkViewDelegate can take the argument with
  the type Context