Remove lzma SDK from .gitignore.
This is the first step to land the LZMA SDK part of Crosswalk
source tree. Taking the files away from .gitignore will make gclient
clean up the source tree of the old extracted files coming from the old
way to fetch the SDK (downloading it and then extracting it).

We should be able to import the SDK in a following commit without
the infra complaining about overwriting untracked files.